refactor(Dashboard): alter structure of Dashboard component for greater clarity separate mobile & dashboard versions in a more readable way move text containing link to eJP into its own file for re-use in 2 places (mobile & desktop versions)
