+++ app/components/pages/Dashboard/index.js
@@ -4,18 +4,23 @@ import { Box, Flex } from 'grid-styled'
 import { th } from '@pubsweet/ui-toolkit'
 
 import media from '../../global/layout/media'
-import SmallParagraph from '../../ui/atoms/SmallParagraph'
-import NativeLink from '../../ui/atoms/NativeLink'
 import StickyFooter from '../../ui/atoms/StickyFooter'
 import Tabs from '../../ui/molecules/Tabs'
 import Submissions from './panels/Submissions'
 import Archived from './panels/Archived'
 import NewSubmissionButton from './NewSubmissionButton'
+import SubmissionsDropdown from './SubmissionsDropdown'
+import EjpLink from './EjpLink'
 
-const BoxHiddenFromMobile = styled(Box)`
+const ContainerHiddenFromMobile = styled(Flex)`
   text-align: right;
   display: none;
   ${media.mobileUp`display: block`};
+`
+
+const MobileOnlyContainer = styled(Box)`
+  display: block;
+  ${media.mobileUp`display: none;`};
 `
 
 const TabbedSubmissions = styled(Tabs)`
@@ -23,22 +28,12 @@ const TabbedSubmissions = styled(Tabs)`
   margin-bottom: ${th('space.3')};
 `
 
-const MobileOnlyStickyFooter = styled(StickyFooter)`
-  display: block;
-  ${media.mobileUp`display: none;`};
-`
-
-const CenterdSmallParagraph = styled(SmallParagraph)`
-  color: ${th('colorTextSecondary')};
-  text-align: center;
-`
 const DashboardPage = ({ history }) => (
-  <Flex>
-    <Box flex="1 1 auto" mx={[0, 0, 0, '16.666%']}>
-      <BoxHiddenFromMobile mb={1} mt={3}>
+  <React.Fragment>
+    <ContainerHiddenFromMobile flex="1 1 auto" mx={[0, 0, 0, '16.666%']}>
+      <Box mb={1} mt={3}>
         <NewSubmissionButton history={history} />
-      </BoxHiddenFromMobile>
-
+      </Box>
       <TabbedSubmissions>
         <Tabs.List>
           <Tabs.Tab>Submissions</Tabs.Tab>
@@ -51,20 +46,16 @@ const DashboardPage = ({ history }) => (
           <Archived />
         </Tabs.Panel>
       </TabbedSubmissions>
+      <EjpLink />
+    </ContainerHiddenFromMobile>
 
-      <CenterdSmallParagraph>
-        Can&#39;t find a submission? You might find it in our full{' '}
-        <NativeLink href="https://submit.elifesciences.org">
-          peer review and submissions
-        </NativeLink>{' '}
-        system
-      </CenterdSmallParagraph>
-    </Box>
-
-    <MobileOnlyStickyFooter>
-      <NewSubmissionButton history={history} />
-    </MobileOnlyStickyFooter>
-  </Flex>
+    <MobileOnlyContainer>
+      <SubmissionsDropdown />
+      <StickyFooter>
+        <NewSubmissionButton history={history} />
+      </StickyFooter>
+    </MobileOnlyContainer>
+  </React.Fragment>
 )
 
 export default DashboardPage
